Update operation handling for mock Certificate Manager DNS authorization
