+++ mockgcp/mockcertificatemanager/service.go
@@ -17,11 +17,12 @@ package mockcertificatemanager
 import (
 	"context"
 	"net/http"
+	"strings"
 
-	"github.com/grpc-ecosystem/grpc-gateway/v2/runtime"
 	"google.golang.org/grpc"
 
 	"github.com/GoogleCloudPlatform/k8s-config-connector/mockgcp/common"
+	"github.com/GoogleCloudPlatform/k8s-config-connector/mockgcp/common/httpmux"
 	"github.com/GoogleCloudPlatform/k8s-config-connector/mockgcp/common/operations"
 	pb "github.com/GoogleCloudPlatform/k8s-config-connector/mockgcp/generated/mockgcp/cloud/certificatemanager/v1"
 	"github.com/GoogleCloudPlatform/k8s-config-connector/mockgcp/pkg/storage"
@@ -57,11 +58,22 @@ func (s *MockService) Register(grpcServer *grpc.Server) {
 }
 
 func (s *MockService) NewHTTPMux(ctx context.Context, conn *grpc.ClientConn) (http.Handler, error) {
-	mux := runtime.NewServeMux()
-
-	if err := pb.RegisterCertificateManagerHandler(ctx, mux, conn); err != nil {
+	mux, err := httpmux.NewServeMux(ctx, conn, httpmux.Options{},
+		pb.RegisterCertificateManagerHandler,
+		s.operations.RegisterOperationsPath("/v1/{prefix=**}/operations/{name}"),
+	)
+	if err != nil {
 		return nil, err
 	}
-
+	mux.RewriteError = func(ctx context.Context, error *httpmux.ErrorResponse) {
+		if error.Code == 404 {
+			if strings.HasPrefix(error.Message, "dnsAuthorization") {
+				error.Message = strings.Replace(error.Message, "dnsAuthorization", "Resource", 1)
+				error.Message = strings.Replace(error.Message, `"`, `'`, 2)
+				error.Message = strings.Replace(error.Message, "not found", "was not found", 1)
+			}
+			error.Errors = nil
+		}
+	}
 	return mux, nil
 }
